tests for signal models separated
