+++ tests/operators/models/test_saturation_recovery.py
@@ -1,4 +1,4 @@
-"""Tests for the mono-exponential decay signal model."""
+"""Tests for saturation recovery signal model."""
 
 # Copyright 2024 Physikalisch-Technische Bundesanstalt
 #
@@ -14,43 +14,43 @@
 
 import pytest
 import torch
-from mrpro.operators.models import MonoExponentialDecay
+from mrpro.operators.models import SaturationRecovery
 from tests.conftest import SHAPE_VARIATIONS_SIGNAL_MODELS
 from tests.conftest import create_parameter_tensor_tuples
 
 
 @pytest.mark.parametrize(
-    ('decay_time', 'result'),
+    ('ti', 'result'),
     [
-        (0, 'm0'),  # short decay time
-        (20, '0'),  # long decay time
+        (0, '0'),  # short ti
+        (20, 'm0'),  # long ti
     ],
 )
-def test_mono_exponential_decay(decay_time, result):
-    """Test for mono-exponential decay signal model.
+def test_saturation_recovery(ti, result):
+    """Test for saturation recovery.
 
     Checking that idata output tensor at ti=0 is close to 0. Checking
     that idata output tensor at large ti is close to m0.
     """
-    model = MonoExponentialDecay(decay_time)
-    m0, decay_constant = create_parameter_tensor_tuples()
-    (image,) = model.forward(m0, decay_constant)
+    model = SaturationRecovery(ti)
+    m0, t1 = create_parameter_tensor_tuples()
+    (image,) = model.forward(m0, t1)
 
     zeros = torch.zeros_like(m0)
 
-    # Assert closeness to m0 for short decay_time
+    # Assert closeness to zero for ti=0
     if result == '0':
         torch.testing.assert_close(image[0, ...], zeros)
-    # Assert closeness to 0 for large decay_time
+    # Assert closeness to m0 for large ti
     elif result == 'm0':
         torch.testing.assert_close(image[0, ...], m0)
 
 
 @SHAPE_VARIATIONS_SIGNAL_MODELS
-def test_mono_exponential_decay_shape(parameter_shape, contrast_dim_shape, signal_shape):
+def test_saturation_recovery_shape(parameter_shape, contrast_dim_shape, signal_shape):
     """Test correct signal shapes."""
-    (decay_time,) = create_parameter_tensor_tuples(contrast_dim_shape, number_of_tensors=1)
-    model_op = MonoExponentialDecay(decay_time)
-    m0, decay_constant = create_parameter_tensor_tuples(parameter_shape, number_of_tensors=2)
-    (signal,) = model_op.forward(m0, decay_constant)
+    (ti,) = create_parameter_tensor_tuples(contrast_dim_shape, number_of_tensors=1)
+    model_op = SaturationRecovery(ti)
+    m0, t1 = create_parameter_tensor_tuples(parameter_shape, number_of_tensors=2)
+    (signal,) = model_op.forward(m0, t1)
     assert signal.shape == signal_shape
